+++ tests/integration_test/objects/box-test-folder.js
@@ -2,8 +2,8 @@
 const utils = require('../lib/utils');
 
 
-const createBoxTestFolder = async(client, name = null, parent_folder_id = '0') => {
-	let folder = await client.folders.create(parent_folder_id, name || utils.randomName());
+const createBoxTestFolder = async(client, name = null, parentFolderId = '0') => {
+	let folder = await client.folders.create(parentFolderId, name || utils.randomName());
 	folder.dispose = async function() {
 		await client.folders.delete(folder.id, { recursive: true });
 		await client.folders.deletePermanently(folder.id);