+++ tests/lib/sessions/ccg-session-test.js
@@ -2,8 +2,6 @@
  * @fileoverview Tests for Anonymous Box API Session.
  */
 
-/* global describe, it, before, beforeEach, after, afterEach */
-
 'use strict';
 
 // ------------------------------------------------------------------------------